chore(components): semantics update
