+++ package.json
@@ -4,7 +4,6 @@
     "module": "./index.js",
     "types": "./index.d.ts",
     "dependencies": {
-        "framer-motion": "^7.6.1",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-svg": "^15.1.9",
